+++ src/lib/automaton/utils.rs
@@ -14,8 +14,8 @@ pub fn mut_add_vec(a: &mut [i32], b: &[i32]) {
     }
 }
 
-pub fn mod_vec(a: &[i32], mu: i32) -> Vec<i32> {
-    a.iter().map(|x| x.rem_euclid(mu)).collect()
+pub fn mod_vec(a: &[i32], mu: u32) -> Vec<i32> {
+    a.iter().map(|x| x.rem_euclid(mu as i32)).collect()
 }
 
 pub fn dyck_transitions_to_ltc_transition(
